after merge with upstram

+++ Rakefile
@@ -1,4 +1,6 @@
 require 'rake'
+require 'fileutils'
+require File.join(File.dirname(__FILE__), 'bin', 'yadr', 'vundle')
 
 desc "Hook our dotfiles into system-standard positions."
 task :install => [:submodule_init, :submodules] do
@@ -16,12 +18,18 @@ task :install => [:submodule_init, :submodules] do
   file_operation(Dir.glob('irb/*')) if want_to_install?('irb/pry configs (more colorful)')
   file_operation(Dir.glob('ruby/*')) if want_to_install?('rubygems config (faster/no docs)')
   file_operation(Dir.glob('ctags/*')) if want_to_install?('ctags config (better js/ruby support)')
+  file_operation(Dir.glob('tmux/*')) if want_to_install?('tmux config')
   file_operation(Dir.glob('vimify/*')) if want_to_install?('vimification of command line tools')
-  file_operation(Dir.glob('{vim,vimrc}')) if want_to_install?('vim configuration (highly recommended)')
+  if want_to_install?('vim configuration (highly recommended)')
+    file_operation(Dir.glob('{vim,vimrc}')) 
+    Rake::Task["install_vundle"].execute
+  end
 
   Rake::Task["install_prezto"].execute
 
   install_fonts if RUBY_PLATFORM.downcase.include?("darwin")
+
+  install_term_theme if RUBY_PLATFORM.downcase.include?("darwin")
 
   success_msg("installed")
 end
@@ -32,14 +40,16 @@ task :install_prezto do
   end
 end
 
-task :update => [:install] do
+task :update do
+  Rake::Task["vundle_migration"].execute if needs_migration_to_vundle?
+  Rake::Task["install"].execute
   #TODO: for now, we do the same as install. But it would be nice
   #not to clobber zsh files
 end
 
 task :submodule_init do
   unless ENV["SKIP_SUBMODULES"]
-    run %{ git submodule update --init --recursive }
+    run %{ git submodule init }
   end
 end
 
@@ -52,11 +62,52 @@ task :submodules do
 
     run %{
       cd $HOME/.yadr
-      git submodule foreach 'git fetch origin; git checkout master; git reset --hard origin/master; git submodule update --recursive; git clean -dfx'
-      git clean -dfx
+      git submodule update --recursive
+      git clean -df
     }
     puts
   end
+end
+
+desc "Performs migration from pathogen to vundle"
+task :vundle_migration do
+  puts "======================================================"
+  puts "Migrating from pathogen to vundle vim plugin manager. "
+  puts "This will move the old .vim/bundle directory to" 
+  puts ".vim/bundle.old and replacing all your vim plugins with"
+  puts "the standard set of plugins. You will then be able to "
+  puts "manage your vim's plugin configuration by editing the "
+  puts "file .vim/vundles.vim"
+  puts "======================================================"
+
+  Dir.glob(File.join('vim', 'bundle','**')) do |sub_path|
+    run %{git config -f #{File.join('.git', 'config')} --remove-section submodule.#{sub_path}}
+    # `git rm --cached #{sub_path}`
+    FileUtils.rm_rf(File.join('.git', 'modules', sub_path))
+  end
+  FileUtils.mv(File.join('vim','bundle'), File.join('vim', 'bundle.old'))
+end
+
+desc "Runs Vundle installer in a clean vim environment"
+task :install_vundle do
+  puts "======================================================"
+  puts "Installing vundle."
+  puts "The installer will now proceed to run BundleInstall."
+  puts "Due to a bug, the installer may report some errors"
+  puts "when installing the plugin 'syntastic'. Fortunately"
+  puts "Syntastic will install and work properly despite the"
+  puts "errors so please just ignore them and let's hope for"
+  puts "an update that fixes the problem!"
+  puts "======================================================"
+
+  puts ""
+  
+  run %{
+    cd $HOME/.yadr
+    git clone https://github.com/gmarik/vundle.git #{File.join('vim','bundle', 'vundle')}
+  }
+
+  Vundle::update_vundle
 end
 
 task :default => 'install'
@@ -79,17 +130,21 @@ def install_rvm_binstubs
 end
 
 def install_homebrew
-  puts "======================================================"
-  puts "Installing Homebrew, the OSX package manager...If it's"
-  puts "already installed, this will do nothing."
-  puts "======================================================"
-  run %{ruby -e "$(curl -fsSkL raw.github.com/mxcl/homebrew/go)"}
+  run %{which brew}
+  unless $?.success?
+    puts "======================================================"
+    puts "Installing Homebrew, the OSX package manager...If it's"
+    puts "already installed, this will do nothing."
+    puts "======================================================"
+    run %{ruby -e "$(curl -fsSkL raw.github.com/mxcl/homebrew/go)"}
+  end
+
   puts
   puts
   puts "======================================================"
   puts "Installing Homebrew packages...There may be some warnings."
   puts "======================================================"
-  run %{brew install ack ctags git hub}
+  run %{brew install zsh ctags git hub tmux reattach-to-user-namespace the_silver_searcher}
   puts
   puts
 end
@@ -100,6 +155,71 @@ def install_fonts
   puts "======================================================"
   run %{ cp -f $HOME/.yadr/fonts/* $HOME/Library/Fonts }
   puts
+end
+
+def install_term_theme
+  puts "======================================================"
+  puts "Installing iTerm2 solarized theme."
+  puts "======================================================"
+  run %{ /usr/libexec/PlistBuddy -c "Add :'Custom Color Presets':'Solarized Light' dict" ~/Library/Preferences/com.googlecode.iterm2.plist }
+  run %{ /usr/libexec/PlistBuddy -c "Merge 'iTerm2/Solarized Light.itermcolors' :'Custom Color Presets':'Solarized Light'" ~/Library/Preferences/com.googlecode.iterm2.plist }
+  run %{ /usr/libexec/PlistBuddy -c "Add :'Custom Color Presets':'Solarized Dark' dict" ~/Library/Preferences/com.googlecode.iterm2.plist }
+  run %{ /usr/libexec/PlistBuddy -c "Merge 'iTerm2/Solarized Dark.itermcolors' :'Custom Color Presets':'Solarized Dark'" ~/Library/Preferences/com.googlecode.iterm2.plist }
+
+  # If iTerm2 is not installed or has never run, we can't autoinstall the profile since the plist is not there
+  if !File.exists?(File.join(ENV['HOME'], '/Library/Preferences/com.googlecode.iterm2.plist'))
+    puts "======================================================"
+    puts "To make sure your profile is using the solarized theme"
+    puts "Please check your settings under:"
+    puts "Preferences> Profiles> [your profile]> Colors> Load Preset.."
+    puts "======================================================"
+    return
+  end
+
+  # Ask the user which theme he wants to install
+  message = "Which theme would you like to apply to your iTerm2 profile?"
+  color_scheme = ask message, iTerm_available_themes
+  color_scheme_file = File.join('iTerm2', "#{color_scheme}.itermcolors")
+
+  # Ask the user on which profile he wants to install the theme
+  profiles = iTerm_profile_list
+  message = "I've found #{profiles.size} #{profiles.size>1 ? 'profiles': 'profile'} on your iTerm2 configuration, which one would you like to apply the Solarized theme to?"
+  profiles << 'All'
+  selected = ask message, profiles
+  
+  if selected == 'All'
+    (profiles.size-1).times { |idx| apply_theme_to_iterm_profile_idx idx, color_scheme_file }
+  else
+    apply_theme_to_iterm_profile_idx profiles.index(selected), color_scheme_file
+  end
+end
+
+def iTerm_available_themes
+   Dir['iTerm2/*.itermcolors'].map { |value| File.basename(value, '.itermcolors')}
+end
+
+def iTerm_profile_list
+  profiles=Array.new
+  begin
+    profiles <<  %x{ /usr/libexec/PlistBuddy -c "Print :'New Bookmarks':#{profiles.size}:Name" ~/Library/Preferences/com.googlecode.iterm2.plist 2>/dev/null}
+  end while $?.exitstatus==0
+  profiles.pop
+  profiles
+end
+
+def ask(message, values)
+  puts message
+  while true
+    values.each_with_index { |val, idx| puts " #{idx+1}. #{val}" }
+    selection = STDIN.gets.chomp
+    if (Float(selection)==nil rescue true) || selection.to_i < 0 || selection.to_i > values.size+1
+      puts "ERROR: Invalid selection.\n\n"
+    else
+      break
+    end
+  end 
+  selection = selection.to_i-1
+  values[selection]
 end
 
 def install_prezto
@@ -122,10 +242,17 @@ def install_prezto
   run %{ mkdir -p $HOME/.zsh.before }
   run %{ mkdir -p $HOME/.zsh.after }
   run %{ mkdir -p $HOME/.zsh.prompts }
+
+  if ENV["SHELL"].include? 'zsh' then
+    puts "Zsh is already configured as your shell of choice. Restart your session to load the new settings"
+  else
+    puts "Setting zsh as your default shell"
+    run %{ chsh -s /bin/zsh }
+  end
 end
 
 def want_to_install? (section)
-  if ENV["ask"]=="true"
+  if ENV["ASK"]=="true"
     puts "Would you like to install configuration files for: #{section}? [y]es, [n]o"
     STDIN.gets.chomp == 'y'
   else
@@ -143,7 +270,7 @@ def file_operation(files, method = :symlink)
     puts "Source: #{source}"
     puts "Target: #{target}"
 
-    if File.exists?(target) || File.symlink?(target)
+    if File.exists?(target) && (!File.symlink?(target) || (File.symlink?(target) && File.readlink(target) != source))
       puts "[Overwriting] #{target}...leaving original at #{target}.backup..."
       run %{ mv "$HOME/.#{file}" "$HOME/.#{file}.backup" }
     end
@@ -166,6 +293,25 @@ def file_operation(files, method = :symlink)
     puts "=========================================================="
     puts
   end
+end
+
+def needs_migration_to_vundle?
+  File.exists? File.join('vim', 'bundle', 'tpope-vim-pathogen')
+end
+
+
+def list_vim_submodules
+  result=`git submodule -q foreach 'echo $name"||"\`git remote -v | awk "END{print \\\\\$2}"\`'`.select{ |line| line =~ /^vim.bundle/ }.map{ |line| line.split('||') }
+  Hash[*result.flatten]
+end
+
+def apply_theme_to_iterm_profile_idx(index, color_scheme_path)
+  values = Array.new
+  16.times { |i| values << "Ansi #{i} Color" }
+  values << ['Background Color', 'Bold Color', 'Cursor Color', 'Cursor Text Color', 'Foreground Color', 'Selected Text Color', 'Selection Color']
+  values.flatten.each { |entry| run %{ /usr/libexec/PlistBuddy -c "Delete :'New Bookmarks':#{index}:'#{entry}'" ~/Library/Preferences/com.googlecode.iterm2.plist } }
+
+  run %{ /usr/libexec/PlistBuddy -c "Merge '#{color_scheme_path}' :'New Bookmarks':#{index}" ~/Library/Preferences/com.googlecode.iterm2.plist }
 end
 
 def success_msg(action)
